темы

+++ main.py
@@ -10,6 +10,10 @@ MAP_FILE = "map.png"
 
 
 class GameView(arcade.Window):
+
+    def __init__(self, width, height, title):
+        super().__init__(width, height, title)
+        self.theme = "light" 
     def setup(self):
         self.get_image()
 
@@ -26,13 +30,25 @@ class GameView(arcade.Window):
             ),
         )
 
+    def on_key_press(self, key, modifiers):
+        if key == arcade.key.T:
+            if self.theme == "light":
+                self.theme = "dark"
+            else:
+                self.theme = "light"
+
+            print(f"Текущая тема: {self.theme}")
+            self.get_image()
+
     def get_image(self):
         server_address = 'https://static-maps.yandex.ru/v1?'
         api_key = 'f3a0fe3a-b07e-4840-a1da-06f18b2ddf13'
         ll_spn = 'll=37.530887,55.703118&spn=0.002,0.002'
-        # Готовим запрос.
 
-        map_request = f"{server_address}{ll_spn}&apikey={api_key}"
+        # Добавляем параметр темы
+        theme_param = f"&theme={self.theme}"
+
+        map_request = f"{server_address}{ll_spn}&apikey={api_key}{theme_param}"
         response = requests.get(map_request)
 
         if not response:
@@ -41,7 +57,6 @@ class GameView(arcade.Window):
             print("Http статус:", response.status_code, "(", response.reason, ")")
             sys.exit(1)
 
-        # Запишем полученное изображение в файл.
         with open(MAP_FILE, "wb") as file:
             file.write(response.content)
 
@@ -52,8 +67,9 @@ def main():
     gameview = GameView(WINDOW_WIDTH, WINDOW_HEIGHT, WINDOW_TITLE)
     gameview.setup()
     arcade.run()
-    # Удаляем за собой файл с изображением.
-    os.remove(MAP_FILE)
+
+    if os.path.exists(MAP_FILE):
+        os.remove(MAP_FILE)
 
 
 if __name__ == "__main__":
